feat: extract Core Session (session.ts) to @opencode-ai/state
- Moved BackgroundJob, EventV2Bridge, InstanceState, Snapshot, Permission,
  RuntimeFlags service types/interfaces from opencode to @opencode-ai/core
- Created minimal interfaces in core for each (just what session.ts needs)
- Moved session.ts (1119 lines) to @opencode-ai/state
- All 6 @/-imports resolved to @opencode-ai/core equivalents
- state/session.ts now has zero @/-imports (fully decoupled)
- opencode session.ts is now a thin re-export wrapper

Packages affected: core, state, opencode
